Added confirmation popups for delete buttons

+++ app/static/js/editor/editor.relationships.mgmt.js
@@ -28,8 +28,10 @@ class RelationshipMgmt {
           obj['section'] = this._section;
           this._source.addRelationship(obj);
         } else if (target.classList.contains('del-rel')) {
-          var rel_id = parseInt(target.getAttribute('data-rel-id'));
-          this._source.deleteRelationship(this._section, rel_id);
+          if (window.confirm('This will delete this relationship! Are you sure?')) {
+            var rel_id = parseInt(target.getAttribute('data-rel-id'));
+            this._source.deleteRelationship(this._section, rel_id);
+          }
         }
         return;
       default:
